set up dummy endpoints with no db
