+++ apps/backend/src/handlers/get-by-id.mjs
@@ -1,16 +1,7 @@
 // Create clients and set shared const values outside of the handler.
 
-// Create a DocumentClient that represents the query to add an item
-import { DynamoDBClient } from '@aws-sdk/client-dynamodb';
-import { DynamoDBDocumentClient, GetCommand } from '@aws-sdk/lib-dynamodb';
-const client = new DynamoDBClient({});
-const ddbDocClient = DynamoDBDocumentClient.from(client);
-
-// Get the DynamoDB table name from environment variables
-const tableName = process.env.SAMPLE_TABLE;
-
 /**
- * A simple example includes a HTTP get method to get one item by id from a DynamoDB table.
+ * A simple example includes a HTTP get method to get one item by id from a dummy nonexistent table.
  */
 export const getByIdHandler = async (event) => {
   if (event.httpMethod !== 'GET') {
@@ -22,23 +13,9 @@ export const getByIdHandler = async (event) => {
   // Get id from pathParameters from APIGateway because of `/{id}` at template.yaml
   const id = event.pathParameters.id;
  
-  // Get the item from the table
-  // https://docs.aws.amazon.com/AWSJavaScriptSDK/latest/AWS/DynamoDB/DocumentClient.html#get-property
-  var params = {
-    TableName : tableName,
-    Key: { id: id },
-  };
-
-  try {
-    const data = await ddbDocClient.send(new GetCommand(params));
-    var item = data.Item;
-  } catch (err) {
-    console.log("Error", err);
-  }
- 
   const response = {
     statusCode: 200,
-    body: JSON.stringify(item)
+    body: JSON.stringify(id)
   };
  
   // All log statements are written to CloudWatch
